moving flow feature to runtime package
