+++ api/src/main/java/com/fnproject/fn/api/RuntimeFeature.java
@@ -1,7 +1,6 @@
 package com.fnproject.fn.api;
 
 /**
- *
  * RuntimeFeatures are classes that configure the Fn Runtime prior to startup and can be loaded by annotating the function class with a {@link FnFeature} annotation
  * Created on 10/09/2018.
  * <p>
@@ -11,7 +10,8 @@ public interface RuntimeFeature {
 
     /**
      * Initialize the runtime context for this function
+     *
      * @param context a runtime context to initalize
      */
-    void initialize(RuntimeContext context) throws Exception;
+    void initialize(RuntimeContext context);
 }
