+++ examples/async-thumbnails/src/main/java/com/fnproject/fn/examples/ThumbnailsFunction.java
@@ -4,7 +4,7 @@ import com.fnproject.fn.api.FnFeature;
 import com.fnproject.fn.api.Headers;
 import com.fnproject.fn.api.RuntimeContext;
 import com.fnproject.fn.api.flow.Flow;
-import com.fnproject.fn.api.flow.FlowFeature;
+import com.fnproject.fn.runtime.flow.FlowFeature;
 import com.fnproject.fn.api.flow.Flows;
 import com.fnproject.fn.api.flow.HttpMethod;
 import io.minio.MinioClient;
