+++ integration-tests/main/test-1-jdk8/src/main/java/com/fnproject/fn/integration/test_1/CompleterFunction.java
@@ -2,7 +2,7 @@ package com.fnproject.fn.integration.test_1;
 
 import com.fnproject.fn.api.FnFeature;
 import com.fnproject.fn.api.flow.Flow;
-import com.fnproject.fn.api.flow.FlowFeature;
+import com.fnproject.fn.runtime.flow.FlowFeature;
 import com.fnproject.fn.api.flow.Flows;
 
 @FnFeature(FlowFeature.class)
